Move Db-Related functionality from server project into new "data" project
