+++ src/TauStellwerk.Data/Dao/ITurnoutDao.cs
@@ -5,9 +5,9 @@
 
 using FluentResults;
 using TauStellwerk.Base;
-using TauStellwerk.Server.Data.Model;
+using TauStellwerk.Data.Model;
 
-namespace TauStellwerk.Server.Data.Dao;
+namespace TauStellwerk.Data.Dao;
 
 public interface ITurnoutDao
 {
