+++ src/TauStellwerk.Data/ImageService/ImageOptions.cs
