+++ src/TauStellwerk.Data/Model/Turnout.cs
@@ -7,7 +7,7 @@ using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using TauStellwerk.Base;
 
-namespace TauStellwerk.Server.Data.Model;
+namespace TauStellwerk.Data.Model;
 
 public class Turnout
 {
