+++ src/TauStellwerk.Server/CommandStations/CommandStationBase.cs
@@ -7,8 +7,8 @@
 
 using FluentResults;
 using TauStellwerk.Base;
-using TauStellwerk.Server.Data;
-using TauStellwerk.Server.Data.Model;
+using TauStellwerk.Data;
+using TauStellwerk.Data.Model;
 
 namespace TauStellwerk.Server.CommandStations;
