+++ src/TauStellwerk.Server/CommandStations/DccEx/DccExSerialCommandStation.cs
@@ -6,7 +6,7 @@
 using System.IO.Ports;
 using FluentResults;
 using TauStellwerk.Base;
-using TauStellwerk.Server.Data.Model;
+using TauStellwerk.Data.Model;
 
 namespace TauStellwerk.Server.CommandStations;
