+++ src/TauStellwerk.Server/CommandStations/ECoS/ECoSCommandStation.cs
@@ -6,8 +6,8 @@
 using System.Net;
 using Microsoft.EntityFrameworkCore;
 using TauStellwerk.Base;
-using TauStellwerk.Server.Data;
-using TauStellwerk.Server.Data.Model;
+using TauStellwerk.Data;
+using TauStellwerk.Data.Model;
 
 namespace TauStellwerk.Server.CommandStations;
