+++ src/TauStellwerk.Server/CommandStations/NullCommandStation.cs
@@ -5,7 +5,7 @@
 
 using FluentResults;
 using TauStellwerk.Base;
-using TauStellwerk.Server.Data.Model;
+using TauStellwerk.Data.Model;
 
 namespace TauStellwerk.Server.CommandStations;
