+++ src/TauStellwerk.Server/Hub/TauHubEngine.cs
@@ -8,7 +8,7 @@ using FluentResults;
 using FluentResults.Extensions;
 using Microsoft.AspNetCore.Mvc;
 using TauStellwerk.Base;
-using TauStellwerk.Server.Data.Dao;
+using TauStellwerk.Data.Dao;
 
 namespace TauStellwerk.Server.Hub;
