+++ src/TauStellwerk.Server/Services/EngineControlService/IEngineControlService.cs
@@ -7,7 +7,7 @@
 
 using FluentResults;
 using TauStellwerk.Base;
-using TauStellwerk.Server.Data.Model;
+using TauStellwerk.Data.Model;
 
 namespace TauStellwerk.Server.Services;
