+++ src/TauStellwerk.Server/Services/TurnoutControlService.cs
@@ -6,8 +6,8 @@
 using System.Collections;
 using FluentResults;
 using TauStellwerk.Base;
+using TauStellwerk.Data.Model;
 using TauStellwerk.Server.CommandStations;
-using TauStellwerk.Server.Data.Model;
 
 namespace TauStellwerk.Server.Services;
