+++ src/TauStellwerk.Tools/CreateTestDb/CreateTestDbTool.cs
@@ -4,8 +4,8 @@
 // </copyright>
 
 using Microsoft.EntityFrameworkCore;
-using TauStellwerk.Server.Data;
-using TauStellwerk.Server.Data.Model;
+using TauStellwerk.Data;
+using TauStellwerk.Data.Model;
 
 namespace TauStellwerk.Tools.CreateTestDb;
