+++ test/TauStellwerk.Server.Test/Services/EngineControlService/LoggingCommandStation.cs
@@ -4,8 +4,8 @@
 // </copyright>
 
 using TauStellwerk.Base;
+using TauStellwerk.Data.Model;
 using TauStellwerk.Server.CommandStations;
-using TauStellwerk.Server.Data.Model;
 
 namespace TauStellwerk.Test.Services.EngineControlService;
